ressources and more ressources animation stuff

+++ game/gameanimationwalk.cpp
@@ -13,7 +13,7 @@
 #include "game/gameanimationfactory.h"
 
 GameAnimationWalk::GameAnimationWalk(Unit* pUnit, QVector<QPoint> movePath)
-    : GameAnimation(GameMap::frameTime),
+    : GameAnimation(static_cast<quint32>(GameMap::frameTime)),
       m_pUnit(pUnit),
       m_movePath(movePath)
 {
@@ -127,7 +127,7 @@ void GameAnimationWalk::loadSprite(QString spriteID, bool addPlayerColor, float 
         if (addPlayerColor)
         {
             QColor color = m_pUnit->getOwner()->getColor();
-            oxygine::Sprite::TweenColor tweenColor(oxygine::Color(color.red(), color.green(), color.blue(), 255));
+            oxygine::Sprite::TweenColor tweenColor(oxygine::Color(static_cast<quint8>(color.red()), static_cast<quint8>(color.green()), static_cast<quint8>(color.blue()), 255));
             oxygine::spTween tween = oxygine::createTween(tweenColor, 1);
             pSprite->addTween(tween);
         }
